move PostActions out of PostContent
